api/requirements.txt
alembic==1.0.3
apscheduler==3.5.3
babel==2.6.0
bcrypt==3.1.4
beautifulsoup4==4.6.3
dateparser==0.7.0
Flask==1.0.2
flask_cors==3.0.7
flask-login==0.4.1
flask_sqlalchemy==2.3.2
flask_script==2.0.6
gspread==3.0.1
inflect==2.1.0
mailjet_rest==1.3.0
newspaper3k==0.2.8
nltk==3.4
oauth2client==4.1.3
Pillow
PostgreSQL-Audit==0.10.0
psycopg2-binary==2.7.6.1
pytest==4.0.0
pytest-flask==0.14.0
pytest-mock==1.10.0
python-keystoneclient==3.18.0
python-swiftclient==3.6.0
requests==2.20.1
simplejson==3.16.0
